Fixed invalid default value for SESSION_DOMAIN
As from commit 522cbd4dd735d5341b00cb797c23099e4a6714d9 the default value for
SESSION_DOMAIN is now set to "apc" if it's not supplied. This will lead to
CSRF token validation failures as described on
https://github.com/CachetHQ/Cachet/issues/2741#issuecomment-528282078

With this commit the SESSION_DOMAIN will just be empty when not explicitly
configured.